[login-Rename] 유저 정보를 가져오는 context를 hooks으로 생성해 분리함 -> useAuth()를 사용하는 컴포넌트에서 경로변경
